Moving AnimatedBackground to demo games
